update sqlitecpp sources to use latest version from git so that it would compile properly.
Also note that to compile sqlite.c in VS2015 the file
has to be selected, and in the file's options under
C/C++ -> Advanced -> Compile As : Compile as C Code /TC
needs to be selected. This for both the debug and the
Release Configuration.

Note further that the .props file holds some of the
preprocessor settings that SQLiteCpp suggests using
for Visual Studio, but these appear to be optional.

that's github repo https://github.com/SRombauts/SQLiteCpp.git
at hash: af78d59349fd013341746b345961e3f19ecae60d